Update README.md
I am more a problem solver than a programmer i like to think. I do know how to program and enjoy it, but my strong suit lies in seeing solutions to baffling problems then implementing them with code, or programming, which to me is nothing more than a tool to this.